* formating changes * initial multiprocess support * bootstrap code in C * remove dead/comented code * improve physical memory allocator * fix heap allocator bugs * fix fat32 driver bugs * fat32 mkdir support * sata driver write support * multi process improvement * shared_ptr bug * UUID calculation changes

+++ libc/string.c
@@ -132,7 +132,7 @@ char *strcat(char *dst, const char *src)
 
 const char *strrchar(const char *s1, int character)
 {
-    for (size_t len = strlen(s1); len >= 0; len--)
+    for (int64_t len = strlen(s1); len >= 0; len--)
     {
         if (s1[len] == character)
         {
